<!-----------------------------------------------------------------------------
Thank you for contributing to the PrestaShop project!

Please take the time to edit the "Answers" rows below with the necessary information.

Check out our contribution guidelines to find out how to complete it:
https://devdocs.prestashop-project.org/8/contribute/contribution-guidelines/#pull-requests

For type and category see:
https://devdocs.prestashop-project.org/8/contribute/contribution-guidelines/pull-requests/#type--category
------------------------------------------------------------------------------>

| Questions         | Answers
| ----------------- | -------------------------------------------------------
| Branch?           | 8.1.x
| Description?      | Optimize the feature values fetching in product page to reduce the number of requests and the memory usage
| Type?             | improvement
| Category?         | BO
| BC breaks?        | no
| Deprecations?     | no
| How to test?      | ~
| UI Tests          | https://github.com/jolelievre/ga.tests.ui.pr/actions/runs/7084101479
| Fixed issue or discussion?     | ~
| Related PRs       | ~
| Sponsor company   | ~

## Condition of tests

I used this module to generate the data in which to improvements have been added:
https://github.com/PrestaShop/PsFixturesCreator/pull/4
https://github.com/PrestaShop/PsFixturesCreator/pull/5

Then used the command from the module to generate a product with a lot of features:
```bash
./bin/console prestashop:product-creator --productsWithCombinations=1 --features=12 --featureValues=12
./bin/console prestashop:product-creator --productsWithCombinations=1 --features=65 --featureValues=25
```

Then the product page in BO was tested (only the new product page from 8.1, not the legacy one) and Blackfire was used to extract some metrics displayed in the following tables.

### Tests in prod mode
| Metric | Number of feature values | Without optimization | With optimization |
| ------ | -------------------------- | --------------------- | ----------------- |
| Time | 12x12 | 6.51s | 5.91s |
| Memory peak usage |   | 73.6Mb | 67.1Mb |
| SQL Queries |   | 497 ms / 2 354 rq | 322 ms / 207 rq |
| Time | 2 but feature has 1000 values | 2.4s | 1.88s |
| Memory peak usage |   | 35.4Mb | 30.7Mb |
| SQL Queries |   | 630 ms / 2 201 rq | 67.7 ms / 197 rq |
| Time | 65x25 | 1m15s | 48.7s |
| Memory peak usage |   | 594Mb | 466Mb |
| SQL Queries |   | 11.9 s / 45 694 rq | 337 ms / 325 rq |

### Tests in dev mode
In dev mode the metrics explode (especially the memory usage) because many things are profiles by Symfony for debug toolbar probably and it prevents using Blackfire for monitoring

| Metric | Number of feature values | Without optimization | With optimization |
| ------ | -------------------------- | --------------------- | ----------------- |
| Time | 12x12 | 26.3s | 20.2s |
| Memory peak usage |   | 306Mb | 281Mb |
| SQL Queries |   | 3.18 s / 2 354 rq | 378 ms / 219 rq |
| Time | 65x25 | No blackfire | No blackfire |
| Memory peak usage |   | No blackfire | No blackfire |
| SQL Queries |   | No blackfire | No blackfire |

## Conclusion

Even though these optimizations improve the performance a bit they are far from enough to handle a product with hundreds o even thousands of features The number of queries and memory related to all this data is one thing, but it seems most of the memory usage comes from having hundreds/thousands of small form types and also the twig rendering all those form types

I advise reworking this feature with an ajax component instead, similar to what as done for combinations, I don't know if pagination would be relevant or not though Maybe the list of features could still be fetched along with the form just for display, maybe not a good idea maybe even the display should be fetched by ajax But the edition itself anyway MUST be handled by ajax so we don't have thousands of inputs in the page, because regardless of the tremendous cost on server side there's even a huge cost for the browser on the client side