@Hlavtox 2023-11-17
@jolelievre Features and feature values must be ordered by name, like on the old page. 👍