@jolelievre 2023-11-17
> @jolelievre Features and feature values must be ordered by name, like on the old page. 👍

In the new product page the features and feature values are ordered by position, this was added on purpose If we want to change this functional behaviour I let you talk about it with @PrestaShop/product-council but that's not the purpose of this issue/PR so I prefer to focus on optimization for now