@Hlavtox 2023-11-17
@jolelievre Feature values don't have a position, they must be sorted alpkabetically. 
`>orderBy('f.position, fv.id_feature_value', 'ASC')` this basically sorts them by ID, because the position of the parent feature is still the same.