@jolelievre 2023-11-17
> @jolelievre Feature values don't have a position, they must be sorted alphabetically. `>orderBy('f.position, fv.id_feature_value', 'ASC')` this basically sorts them by ID, because the position of the parent feature is still the same.

I believe this is mostly to "group" the returned values by feature parent and keep a consistent return for expected tests results, but I'll try to add a sort in the choice provider instead