@Hlavtox 2023-11-17
Remember there can be thousands of feature values returned, why not just add `ORDER BY fvl.value`?