@MatShir 2023-11-17
Regarding the sort, I’m going back to my statement for the order by position in the search input. I agree, it can be very inconvenient to look for data since the feature can be a very long list, with feature of different for different product type. So the features mix feature for example for TV and T-shirt, so the position sorting making it not really relevant.

To conclude, ideally feature and feature values should be sort by alphabetic order, but I don’t know if it fits the PR scope. I would be a nice to have