@MatShir 2023-11-17
The sorting by position could be more interesting to order the features after the save. So the merchant would have a preview of how the list would be in the FO but keep the alpha order when looking for feature and feature value for an easy search. Just an improv idea