@Hlavtox 2023-11-21
@jolelievre Works perfectly! Last thing we should optimize is this - fetch the lang values directly.

![Snímek obrazovky 2023-11-21 115714](https://github.com/PrestaShop/PrestaShop/assets/6097524/ffa9a37f-a9d1-4ba0-af7c-5f56d53e0cae)