@jolelievre 2023-11-22
@Hlavtox this one is trickier because it NEED the multi lang values, so the pagination can't work But I'll try and see if I can create yet another method (or modify the actual one) that doesn't use pagination and thus would allow doing a single request