@paulnoelcholot 2023-12-04
Hello @jolelievre,

I tested your PR and it's GOOD for me! :tada: 

Thanks!